+++ Makefile
@@ -4,17 +4,20 @@ os-image : boot_sect.bin kernel.bin interrupts.bin
 boot_sect.bin : boot/bootloader.asm
 	nasm -I 'boot' boot/bootloader.asm -f bin -o boot_sect.bin
 	
-kernel.bin : kernel_entry.o kernel.o usermode.o shell.o isr.o syscall.o screen.o keyboard.o interrupts.o idt.o I_O_asm_helpers.o utils.o
-	ld -m elf_i386 -o kernel.bin -Ttext 0x1000 kernel_entry.o kernel.o usermode.o shell.o isr.o syscall.o screen.o keyboard.o interrupts.o idt.o I_O_asm_helpers.o utils.o --oformat binary
+kernel.bin : kernel_entry.o kernel.o usermode.o shell.o isr.o syscall.o screen.o keyboard.o interrupts.o idt.o IO.o mem.o stdlib.o
+	ld -m elf_i386 -o kernel.bin -Ttext 0x1000 kernel_entry.o kernel.o usermode.o shell.o isr.o syscall.o screen.o keyboard.o interrupts.o idt.o IO.o mem.o stdlib.o --oformat binary
 
-interrupts.bin : interrupts.o isr.o syscall.o keyboard.o idt.o screen.o I_O_asm_helpers.o utils.o
-	ld -m elf_i386 -o interrupts.bin -Ttext 0x1000 interrupts.o isr.o syscall.o keyboard.o idt.o screen.o I_O_asm_helpers.o utils.o --oformat binary	
+interrupts.bin : interrupts.o isr.o syscall.o keyboard.o idt.o screen.o IO.o mem.o stdlib.o
+	ld -m elf_i386 -o interrupts.bin -Ttext 0x1000 interrupts.o isr.o syscall.o keyboard.o idt.o screen.o IO.o mem.o stdlib.o --oformat binary	
 
 kernel_entry.o : boot/kernel_entry.asm
 	nasm boot/kernel_entry.asm -f elf32 -o kernel_entry.o
 
 shell.o : shell/shell.c
-	gcc -g -fno-pie -ffreestanding -m32 -c shell/shell.c -o shell.o 	
+	gcc -g -fno-pie -ffreestanding -m32 -c shell/shell.c -o shell.o 
+
+stdlib.o : stdlib/stdlib.c
+	gcc -g -fno-pie -ffreestanding -m32 -c stdlib/stdlib.c -o stdlib.o 	
 
 interrupts.o : interrupts/interrupts.asm
 	nasm interrupts/interrupts.asm -f elf32 -o interrupts.o
@@ -22,11 +25,11 @@ interrupts.o : interrupts/interrupts.asm
 usermode.o : kernel/usermode.asm
 	nasm kernel/usermode.asm -f elf32 -o usermode.o
 
-utils.o : kernel/utils.c
-	gcc -g -fno-pie -ffreestanding -m32 -c kernel/utils.c -o utils.o
+mem.o : kernel/mem.c
+	gcc -g -fno-pie -ffreestanding -m32 -c kernel/mem.c -o mem.o
 
-I_O_asm_helpers.o : kernel/I_O_asm_helpers.c
-	gcc -g -fno-pie -ffreestanding -m32 -c kernel/I_O_asm_helpers.c -o I_O_asm_helpers.o
+IO.o : kernel/IO.c
+	gcc -g -fno-pie -ffreestanding -m32 -c kernel/IO.c -o IO.o
 
 syscall.o : drivers/syscall.c
 	gcc -g -fno-pie -ffreestanding -m32 -c drivers/syscall.c -o syscall.o
