+++ kernel/kernel.c
@@ -1,9 +1,7 @@
 #include "../drivers/screen.h"
 #include "../drivers/keyboard.h"
-#include "../interrupts//isr.h"
-#include "../shell/shell.h"
+#include "../interrupts/isr.h"
 #include "../drivers/syscall.h"
-#include "utils.h"
 
 // enterUserMode is our wrapper assembly function that jumps to user mode and then
 // starts the shell
